fetch_take_home/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
requests-mock==1.11.0
pytest-mock==3.12.0